get basics working with v0.2
