+++ Project.toml
@@ -1,12 +1,13 @@
 name = "AutoBZ"
 uuid = "00d1ae05-4f20-4598-b864-bbb2ed8900e6"
 authors = ["lxvm <lorenzo@vanmunoz.com> and contributors"]
-version = "0.1.0"
+version = "0.2.0"
 
 [deps]
 AutoBZCore = "66bd3e16-1600-45cf-8f55-0b550710682b"
 BaryRational = "91aaffc3-5777-4842-85b7-5d3d5d6a3494"
 EquiBaryInterp = "c5458111-f826-47c0-abe9-f36aa00852ec"
+FourierSeriesEvaluators = "2a892dea-6eef-4bb5-9d1c-de966c9f6db5"
 HChebInterp = "78faba9b-a54b-441f-8118-62407cbe4e59"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -14,23 +15,20 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
-AutoBZCore = "0.1.6"
+AutoBZCore = "0.2"
 BaryRational = "0.1"
 EquiBaryInterp = "0.1"
 HChebInterp = "0.1"
 Reexport = "1"
-Requires = "1"
 StaticArrays = "1"
 SymmetryReduceBZ = "0.1"
-julia = "1.6"
+julia = "1.9"
 
 [extensions]
-AuxIntegrandsExt = "AuxQuad"
 BrillouinPlotlyJSExt = ["Brillouin", "PlotlyJS"]
 SymmetryReduceBZExt = ["SymmetryReduceBZ", "Polyhedra"]
 
 [extras]
-AuxQuad = "b59dd6d0-509b-4777-bfd1-cd8479d7b9e5"
 Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
@@ -41,7 +39,6 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 test = ["SymmetryReduceBZ", "Test"]
 
 [weakdeps]
-AuxQuad = "b59dd6d0-509b-4777-bfd1-cd8479d7b9e5"
 Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
