+++ src/AutoBZ.jl
@@ -9,30 +9,18 @@ define custom BZ integrands.
 """
 module AutoBZ
 
-!isdefined(Base, :get_extension) && using Requires
-@static if !isdefined(Base, :get_extension)
-    function __init__()
-        @require SymmetryReduceBZ = "49a35663-c880-4242-bebb-1ec8c0fa8046" include("../ext/SymmetryReduceBZExt.jl")
-        @require Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0" begin
-            @require PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a" include("../ext/BrillouinPlotlyJSExt.jl")
-        end
-    end
-end
-
 using LinearAlgebra
 using LinearAlgebra: checksquare
 
 using StaticArrays
 using Reexport
+@reexport using FourierSeriesEvaluators
 @reexport using AutoBZCore
 
-import AutoBZCore: SymRep, symmetrize_,
-    Integrand, FourierIntegrand, construct_problem,
-    FourierSeriesEvaluators.period, FourierSeriesEvaluators.contract,
-    FourierSeriesEvaluators.evaluate, FourierSeriesEvaluators.coefficients,
-    FourierSeriesEvaluators.show_details, FourierSeriesEvaluators.deriv,
-    FourierSeriesEvaluators.offset, FourierSeriesEvaluators.shift,
-    AutoSymPTR.npt_update
+import FourierSeriesEvaluators: period, deriv, offset, shift,
+    contract, evaluate, coefficients, show_details
+import AutoBZCore: SymRep, symmetrize_, construct_problem, AutoSymPTR.npt_update
+
 
 using EquiBaryInterp: LocalEquiBaryInterp
 using BaryRational: aaa
